Fix Limit & Offset field types for QueryMembersRequest

+++ Assets/Plugins/StreamChat/Core/DTO/Requests/QueryMembersRequestDTO.cs
@@ -41,7 +41,7 @@ namespace StreamChat.Core.DTO.Requests
         public string Id { get; set; }
 
         [Newtonsoft.Json.JsonProperty("limit", Required = Newtonsoft.Json.Required.Default, NullValueHandling = Newtonsoft.Json.NullValueHandling.Ignore)]
-        public double? Limit { get; set; }
+        public int? Limit { get; set; }
 
         /// <summary>
         /// List of members to search in distinct channels
@@ -50,7 +50,7 @@ namespace StreamChat.Core.DTO.Requests
         public System.Collections.Generic.List<ChannelMemberDTO> Members { get; set; }
 
         [Newtonsoft.Json.JsonProperty("offset", Required = Newtonsoft.Json.Required.Default, NullValueHandling = Newtonsoft.Json.NullValueHandling.Ignore)]
-        public double? Offset { get; set; }
+        public int? Offset { get; set; }
 
         /// <summary>
         /// Array of sort parameters
